Revert "Migrate op to new infra: embedding (#33122)"
This reverts commit a23fcef172d6a0b3f00a0c67b9646412e505717f.
